daemon: Don't try to add admin users to non existing groups
The extra admin groups list is a build-time option, but there is no guarantee
the groups are available on the installed system. Currently if an extra admin
group is missing on the system, accountsservice's will fail to create users
because the underlying useradd call will fail.

This commit fixes the issue, by pre-validating the extra admin groups
list.