NEWS: Drop NEWS file
The NEWS file is an attempt to distill what changes have gone in since
last release from the commits.

But accountsservice is a pretty slow moving project and it doesn't
really provide that much value over reading the commits directly.

This commit drops the file, to make it easier to do releases.