meson: Handle weird git clone directory names better
When generating the version for the project, the first step is
to determine if the source is from a release or a clone from git.

The script does this by looking at the directory name and seeing
if it has a hyphen in it. If so, it assumes it's a tarball.

This is a bad heuristic, since clones from git can be put into
any directory name the cloner wants, including directory names
with hyphens!

This commit fixes that edge case by explicitly looking for a clone
from git first.

https://gitlab.freedesktop.org/accountsservice/accountsservice/-/issues/99