meson: Fix up version generation heursitics
They were all broken before for the tarball case and I didn't
notice because I was checking the tarball in a subdirectory of
the git repo so it was just exercising the git path.

This commit makes one more attempt at fixing the issue, and
also provides some protection against users that does shallow
clones.

https://gitlab.freedesktop.org/accountsservice/accountsservice/-/issues/99