src: Run entire tree through uncrustify
The accountsservice coding style is less than pristine, and it would be
good to improve that going forward.

Its coding style is ostensibly the same as plymouths, though, and
plymouth has an uncrustify config hammered out already.

This commit runs the tree through that config to get things in better
shape.