user: Support new LocalAccount property in cache file
At the moment an admin can decide whether or not a user is a system
account by setting SystemAccount= to true or false in the users
cache file, but there's no way to to do the same sort of configuration
for deciding whether or not a user is a local account.

This commit adds support for a new LocalAccount= key in the cache file.

Note, by default this key won't get written into the cache file and
instead accountsservice will continue to rely on it's "user is in
/etc/shadow" heuristic.

The key only gets rewritten into the file during cache file
serialization if an admin added it there first.

Closes: https://gitlab.freedesktop.org/accountsservice/accountsservice/-/issues/110