subprojects: Add mocklibc to package cache
Many builders don't want to talk to the network at build time, but
we currently do fetch mocklibc from the net.

This commit adds an in repo copy of it it, to resolve the problem.

Note, the project is very slow moving at this point so it's
unlikely we'll need to update it later.

Closes https://gitlab.freedesktop.org/accountsservice/accountsservice/-/issues/111