user-manager: Cope with buggy callers better
The user manager owns the users returned from
act_user_manager_get_user*, but it's possible callers may not realize
that.

If a caller does unref the result while a user fetch is in progress, it
can lead to crashes that are hard to pin back to the caller.

This commit adds a weak reference to the user object while fetch user
requests are in flight, and cancels those requests, if the user object
is unexpectantly destroyed in the meanwhile.