user-manager: Deduplicate ActUser objects when possible
If act_user_manager_get_user_by_id is called twice in a row
for the same user, two ActUser objects get generated, each
ultimately representing the same user.

That is less than ideal since the ActUser objects are owned
by the user manager, not by the callers.

This commit tries to minimize the amount of duplicate ActUser
objects that can get put in the wild, by checking for
existig in-flight requests and consolidating them.

Note there can still be duplicated users if there is a
act_user_manager_get_user and act_user_manager_get_user_by_id call
for the same user at the same time. There's no way to deduplicate
the objects in that case.